# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API Server for Generation Barometer Application",
        "title": "Generation Barometer API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "generationth@generation.org"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/",
    "paths": {
        "/admin/barometer": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get statistics about user barometer data (Admin only)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Get barometer statistics",
                "responses": {
                    "200": {
                        "description": "Barometer statistics retrieved",
                        "schema": {
                            "$ref": "#/definitions/utils.StandardResponse"
                        }
                    },
                    "500": {
                        "description": "Error fetching data",
                        "schema": {
                            "$ref": "#/definitions/utils.StandardResponse"
                        }
                    }
                }
            }
        },
        "/admin/chart-data": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get data for charts (Admin only)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Get chart data",
                "responses": {
                    "200": {
                        "description": "Chart data retrieved",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object",
                                "properties": {
                                    "DailyReflection": {
                                        "type": "integer"
                                    },
                                    "day": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to fetch chart data",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/admin/reflections": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get all user reflections with pagination (Admin only)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Get all reflections",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Items per page",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Reflections retrieved",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "data": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/definitions/models.ReflectionWithUser"
                                    }
                                },
                                "success": {
                                    "type": "boolean"
                                },
                                "total": {
                                    "type": "integer"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to fetch reflections",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                },
                                "message": {
                                    "type": "string"
                                },
                                "success": {
                                    "type": "boolean"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/admin/reflections/chartday": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get daily barometer statistics (Admin only)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Get daily barometer data",
                "parameters": [
                    {
                        "type": "string",
                        "default": "90d",
                        "description": "Time range (90d, 30d, 7d)",
                        "name": "timeRange",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Barometer data retrieved",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.BarometerData"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid time range",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to fetch data",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/admin/userreflections/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get a specific user with all their reflections (Admin only)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Get user with reflections",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User and reflections retrieved",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.StandardResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/services.UserWithReflections"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid user ID",
                        "schema": {
                            "$ref": "#/definitions/utils.StandardResponse"
                        }
                    },
                    "500": {
                        "description": "Error retrieving data",
                        "schema": {
                            "$ref": "#/definitions/utils.StandardResponse"
                        }
                    }
                }
            }
        },
        "/admin/users": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get a list of all users (Admin only)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Get all users",
                "responses": {
                    "200": {
                        "description": "Users retrieved",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.StandardResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/models.User"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "403": {
                        "description": "Access denied",
                        "schema": {
                            "$ref": "#/definitions/utils.StandardResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "models.BarometerData": {
            "type": "object",
            "properties": {
                "Comfort Zone": {
                    "type": "integer"
                },
                "Panic Zone": {
                    "type": "integer"
                },
                "Stretch Zone - Enjoying the Challenges": {
                    "type": "integer"
                },
                "Stretch Zone - Overwhelmed": {
                    "type": "integer"
                },
                "date": {
                    "type": "string"
                }
            }
        },
        "models.Reflection": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "description": "Added for daily reflection limit",
                    "type": "string"
                },
                "date": {
                    "type": "string"
                },
                "day": {
                    "description": "Custom field for day, could be a date or unique identifier",
                    "type": "string"
                },
                "reflection": {
                    "description": "Renamed for clarity",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.ReflectionContent"
                        }
                    ]
                },
                "user_id": {
                    "description": "Reference to the user",
                    "type": "string"
                }
            }
        },
        "models.ReflectionContent": {
            "type": "object",
            "properties": {
                "barometer": {
                    "type": "string"
                },
                "non_tech_sessions": {
                    "$ref": "#/definitions/models.SessionDetails"
                },
                "tech_sessions": {
                    "$ref": "#/definitions/models.SessionDetails"
                }
            }
        },
        "models.ReflectionWithUser": {
            "type": "object",
            "properties": {
                "date": {
                    "type": "string"
                },
                "firstName": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "jsdNumber": {
                    "type": "string"
                },
                "lastName": {
                    "type": "string"
                },
                "reflection": {
                    "type": "object",
                    "properties": {
                        "barometer": {
                            "type": "string"
                        },
                        "nonTechSessions": {
                            "type": "object",
                            "properties": {
                                "happy": {
                                    "type": "string"
                                },
                                "improve": {
                                    "type": "string"
                                },
                                "sessionName": {
                                    "type": "array",
                                    "items": {
                                        "type": "string"
                                    }
                                }
                            }
                        },
                        "techSessions": {
                            "type": "object",
                            "properties": {
                                "happy": {
                                    "type": "string"
                                },
                                "improve": {
                                    "type": "string"
                                },
                                "sessionName": {
                                    "type": "array",
                                    "items": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "models.SessionDetails": {
            "type": "object",
            "properties": {
                "happy": {
                    "type": "string"
                },
                "improve": {
                    "type": "string"
                },
                "session_name": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "models.User": {
            "type": "object",
            "properties": {
                "_id": {
                    "description": "MongoDB generated ID",
                    "type": "string"
                },
                "cohort_number": {
                    "type": "integer"
                },
                "email": {
                    "type": "string"
                },
                "first_name": {
                    "type": "string"
                },
                "jsd_number": {
                    "description": "JSD Number เลขที่นักเรียน",
                    "type": "string"
                },
                "last_name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "reflections": {
                    "description": "This is where the reflections should be",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Reflection"
                    }
                },
                "role": {
                    "description": "Add role field (admin/user)",
                    "type": "string"
                }
            }
        },
        "services.UserWithReflections": {
            "type": "object",
            "properties": {
                "reflections": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Reflection"
                    }
                },
                "user": {
                    "$ref": "#/definitions/models.User"
                }
            }
        },
        "utils.StandardResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Enter your bearer token in the format **Bearer \u003ctoken\u003e**",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}